src/main/java/com/example/travelbooking/service/HotelService.java
package com.example.travelbooking.service;

import com.example.travelbooking.model.Hotel;
import com.example.travelbooking.repository.HotelRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
@AllArgsConstructor
public class HotelService {

    private HotelRepository hotelRepository;

    public List<Hotel> searchHotels(String location) {
        return hotelRepository.findByLocation(location);
    }

    public List<Hotel> getAllHotels() {
        return hotelRepository.findAll();
    }

    public Optional<Hotel> getHotelById(Long id) {
        return hotelRepository.findById(id);
    }

    public Hotel saveHotel(Hotel hotel) {
        return hotelRepository.save(hotel);
    }

    public void deleteHotel(Long id) {
        hotelRepository.deleteById(id);
    }
}
